make sources an array
